+++ src/index.ts
@@ -1,3 +1,11 @@
-export {Server, Connection, IServerEvents, IServerOptions} from './server'
-export {Client, IClientEvents, IClientOptions} from './client'
+export { IServerEvents } from './interface/IServerEvents'
+export { IServerOptions } from './interface/IServerOptions'
+
+export { IClientEvents } from './interface/IClientEvents'
+export { IClientOptions } from './interface/IClientOptions'
+
+export { Connection } from './Connection'
+export { Server } from './Server'
+export { Client } from './Client'
+
 export const version: string = require('../package').version
