+++ src/type/Handler.ts
@@ -1,0 +1,4 @@
+import { Connection } from '../Connection'
+import { Message } from './Message'
+
+export type Handler = (request: Message, connection: Connection) => Promise<Message>
