+++ src/type/Message.ts
@@ -1,0 +1,3 @@
+import * as protobuf from 'protobufjs/minimal'
+
+export type Message = protobuf.Message<{}> | {[k: string]: any}
