+++ test/index.ts
@@ -7,7 +7,7 @@ import * as assert from 'assert'
 import * as path from 'path'
 import * as crypto from 'crypto'
 import {Server, Client} from './../src'
-import * as wsrpc_client from './../src/client'
+import * as wsrpc_client from '../src/Client'
 import {waitForEvent} from './../src/utils'
 import {TestService, TextMessage} from './../protocol/test'
 import * as rpcproto from './../protocol/rpc'
